+++ zingolib/src/lightclient/send.rs
@@ -15,7 +15,6 @@ use zcash_proofs::prover::LocalTxProver;
 use super::{LightClient, LightWalletSendProgress};
 use crate::{utils::zatoshis_from_u64, wallet::Pool};
 
-#[cfg(feature = "zip317")]
 use {
     crate::{error::ZingoLibError, wallet::tx_map_and_maybe_trees::TxMapAndMaybeTrees},
     std::{num::NonZeroU32, ops::DerefMut},
@@ -26,40 +25,6 @@ use {
     zingoconfig::ChainType,
 };
 
-#[cfg(feature = "zip317")]
-type GISKit = GreedyInputSelector<
-    TxMapAndMaybeTrees,
-    zcash_client_backend::fees::zip317::SingleOutputChangeStrategy,
->;
-
-/// converts from raw receivers to TransactionRequest
-pub fn receivers_becomes_transaction_request(
-    receivers: Vec<(Address, NonNegativeAmount, Option<MemoBytes>)>,
-) -> Result<TransactionRequest, zcash_client_backend::zip321::Zip321Error> {
-    let mut payments = vec![];
-    for out in receivers.clone() {
-        payments.push(Payment {
-            recipient_address: out.0,
-            amount: out.1,
-            memo: out.2,
-            label: None,
-            message: None,
-            other_params: vec![],
-        });
-    }
-
-    TransactionRequest::new(payments)
-}
-
-use thiserror::Error;
-#[derive(Debug, Error)]
-pub enum DoProposeError {
-    #[error("{0}")]
-    Receiver(zcash_client_backend::zip321::Zip321Error),
-    #[error("{0}")]
-    Proposal(crate::wallet::tx_map_and_maybe_trees::TxMapAndMaybeTreesError),
-}
-
 impl LightClient {
     async fn get_submission_height(&self) -> Result<BlockHeight, String> {
         Ok(BlockHeight::from_u32(
@@ -67,85 +32,6 @@ impl LightClient {
                 .await?
                 .height as u32,
         ) + 1)
-    }
-
-    /// Unstable function to expose the zip317 interface for development
-    // TOdo: add correct functionality and doc comments / tests
-    // TODO: Add migrate_sapling_to_orchard argument
-    #[cfg(feature = "zip317")]
-    pub async fn do_propose_spend(
-        &self,
-        receivers: Vec<(Address, NonNegativeAmount, Option<MemoBytes>)>,
-    ) -> Result<crate::data::proposal::TransferProposal, String> {
-        let request =
-            receivers_becomes_transaction_request(receivers).map_err(|e| e.to_string())?;
-
-        let change_strategy = zcash_client_backend::fees::zip317::SingleOutputChangeStrategy::new(
-            zcash_primitives::transaction::fees::zip317::FeeRule::standard(),
-            None,
-            ShieldedProtocol::Orchard,
-        ); // review consider change strategy!
-
-        let input_selector = GISKit::new(
-            change_strategy,
-            zcash_client_backend::fees::DustOutputPolicy::default(),
-        );
-
-        let mut tmamt = self
-            .wallet
-            .transaction_context
-            .transaction_metadata_set
-            .write()
-            .await;
-
-        let proposal = zcash_client_backend::data_api::wallet::propose_transfer::<
-            TxMapAndMaybeTrees,
-            ChainType,
-            GISKit,
-            ZingoLibError,
-        >(
-            tmamt.deref_mut(),
-            &self.wallet.transaction_context.config.chain,
-            zcash_primitives::zip32::AccountId::ZERO,
-            &input_selector,
-            request,
-            NonZeroU32::MIN, //review! use custom constant?
-        )
-        .map_err(|e| ZingoLibError::Error(format!("error this function todo error {:?}", e)))?;
-
-        let mut latest_proposal_lock = self.latest_proposal.write().await;
-        *latest_proposal_lock = Some(crate::data::proposal::ZingoProposal::Transfer(
-            proposal.clone(),
-        ));
-        Ok(proposal)
-    }
-
-    /// Unstable function to expose the zip317 interface for development
-    // TOdo: add correct functionality and doc comments / tests
-    #[cfg(feature = "zip317")]
-    pub async fn do_propose_shield(
-        &self,
-        _address_amount_memo_tuples: Vec<(&str, u64, Option<MemoBytes>)>,
-    ) -> Result<crate::data::proposal::ShieldProposal, String> {
-        todo!()
-    }
-
-    /// Unstable function to expose the zip317 interface for development
-    // TODO: add correct functionality and doc comments / tests
-    #[cfg(feature = "zip317")]
-    pub async fn do_send_proposal(&self) -> Result<Vec<TxId>, String> {
-        if let Some(proposal) = self.latest_proposal.read().await.as_ref() {
-            match proposal {
-                crate::lightclient::ZingoProposal::Transfer(_) => {
-                    Ok(vec![TxId::from_bytes([1u8; 32])])
-                }
-                crate::lightclient::ZingoProposal::Shield(_) => {
-                    Ok(vec![TxId::from_bytes([222u8; 32])])
-                }
-            }
-        } else {
-            Err("No proposal. Call do_propose first.".to_string())
-        }
     }
 
     /// Send funds
